Main/Backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
sentence-transformers
scikit-learn
numpy
mysql-connector-python
bcrypt
